requirements.txt
pymongo
tikapi
apache-airflow-providers-mongo
